# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  title: string;
  /** ISO date string YYYY-MM-DD */
  eventDate: string;
  /** Human-readable time e.g. "7:00 PM – Midnight" */
  eventTime: string;
  location: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  /** ISO date string YYYY-MM-DD */
  eventDate: string;
  eventTime: string;
  location: string;
  /** @nullable */
  description?: string | null;
}

export interface InstagramPost {
  id: number;
  /** Instagram post shortcode extracted from the URL */
  shortcode: string;
  /**
   * Optional label shown in admin
   * @nullable
   */
  caption: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface CreateInstagramPostBody {
  /** Instagram post shortcode (from URL instagram.com/p/<shortcode>/) */
  shortcode: string;
  /** @nullable */
  caption?: string | null;
  displayOrder: number;
}