# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all events sorted by date ascending
 * @summary List all events
 */
export const ListEventsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  eventDate: zod.string().describe("ISO date string YYYY-MM-DD"),
  eventTime: zod
    .string()
    .describe('Human-readable time e.g. \"7:00 PM – Midnight\"'),
  location: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListEventsResponse = zod.array(ListEventsResponseItem);

/**
 * @summary Create a new event
 */
export const CreateEventBody = zod.object({
  title: zod.string(),
  eventDate: zod.string().describe("ISO date string YYYY-MM-DD"),
  eventTime: zod.string(),
  location: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Delete an event
 */
export const DeleteEventParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List featured Instagram posts
 */
export const ListInstagramPostsResponseItem = zod.object({
  id: zod.number(),
  shortcode: zod
    .string()
    .describe("Instagram post shortcode extracted from the URL"),
  caption: zod.string().nullable().describe("Optional label shown in admin"),
  displayOrder: zod.number(),
  createdAt: zod.string(),
});
export const ListInstagramPostsResponse = zod.array(
  ListInstagramPostsResponseItem,
);

/**
 * @summary Add an Instagram post to the feed
 */
export const CreateInstagramPostBody = zod.object({
  shortcode: zod
    .string()
    .describe(
      "Instagram post shortcode (from URL instagram.com\/p\/<shortcode>\/)",
    ),
  caption: zod.string().nullish(),
  displayOrder: zod.number(),
});

/**
 * @summary Remove an Instagram post from the feed
 */
export const DeleteInstagramPostParams = zod.object({
  id: zod.coerce.number(),
});